+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [0.29.0](https://github.com/firetower-cloud/firetower/compare/firetower-v0.28.2...firetower-v0.29.0) (2026-09-10)
+
+
+### Features
+
+* annotate application previews and send feedback to agents ([1e56ce4](https://github.com/firetower-cloud/firetower/commit/1e56ce4a441f2b76889ca5dab493fdada524c604))
+* annotate application previews and send feedback to agents ([ab2fc2a](https://github.com/firetower-cloud/firetower/commit/ab2fc2aa7cf2d1d4935950c3dde267b86178d299))
+
+
+### Bug Fixes
+
+* take down an approval once it has been answered ([46802f0](https://github.com/firetower-cloud/firetower/commit/46802f0ecd4fa936a34e8f104b16b0c94c6dc51a))
+
+## [0.28.2](https://github.com/firetower-cloud/firetower/compare/firetower-v0.28.1...firetower-v0.28.2) (2026-09-10)
+
+
+### Bug Fixes
+
+* let a session whose agent has gone be started again ([32c4210](https://github.com/firetower-cloud/firetower/commit/32c42101717bf10acf303f2e230b3ece4bee1056))
+* say why a turn failed ([731f21f](https://github.com/firetower-cloud/firetower/commit/731f21f3c14de04e3686755da6f68ed89120ba81))
+
+## [0.28.1](https://github.com/firetower-cloud/firetower/compare/firetower-v0.28.0...firetower-v0.28.1) (2026-09-10)
+
+
+### Bug Fixes
+
+* draw a modal over the window, not inside the rail ([3bc2aba](https://github.com/firetower-cloud/firetower/commit/3bc2aba311f0402c1cebec94522f1fbd95b17292))
+
 ## [0.28.0](https://github.com/firetower-cloud/firetower/compare/firetower-v0.27.1...firetower-v0.28.0) (2026-09-10)